EventBus API in seperate Module
