+++ base/rico-eventbus/rico-eventbus.gradle
@@ -14,13 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.rico.server.remoting.test.eventbus;
 
-import dev.rico.event.Topic;
-
-public interface EventBusTestConstants {
-    String EVENT_BUS_SUBSCIBER_CONTROLLER_NAME = "EventBusSubscriberController";
-    String EVENT_BUS_PUBLISHER_CONTROLLER_NAME = "EventBusPublisherController";
-    String CALL_ACTION = "call";
-    Topic<String> TEST_TOPIC = Topic.create();
+dependencies {
+    compile project(':rico-core')
 }
