+++ base/rico-eventbus/src/test/java/dev/rico/event/TopicTest.java
@@ -1,26 +1,19 @@
-/*
- * Copyright 2018-2019 Karakun AG.
- * Copyright 2015-2018 Canoo Engineering AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package dev.rico.server.remoting.event;
+package dev.rico.event;
 
-import dev.rico.server.remoting.event.Topic;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-public class TestTopic {
+public class TopicTest {
+
+    @Test(expectedExceptions = IllegalArgumentException.class)
+    public void checkNonBlank() {
+        Topic.create("");
+    }
+
+    @Test(expectedExceptions = NullPointerException.class)
+    public void checkNonNull() {
+        Topic.create(null);
+    }
 
     @Test
     public void testUniqueId() {
@@ -45,6 +38,4 @@ public class TestTopic {
         Assert.assertEquals(topicA.getName(), "TopicA");
         Assert.assertEquals(topicB.getName(), "TopicB");
     }
-
-
 }
