+++ remoting/rico-distributed-eventbus/src/main/java/dev/rico/internal/remoting/eventbus/distributed/DistributedEventBus.java
@@ -16,13 +16,13 @@
  */
 package dev.rico.internal.remoting.eventbus.distributed;
 
+import dev.rico.event.MessageEventContext;
+import dev.rico.event.MessageListener;
+import dev.rico.event.Topic;
 import dev.rico.internal.server.remoting.event.AbstractEventBus;
 import dev.rico.internal.server.remoting.event.MessageEventImpl;
 import dev.rico.internal.core.Assert;
 import dev.rico.core.functional.Subscription;
-import dev.rico.server.remoting.event.MessageEventContext;
-import dev.rico.server.remoting.event.MessageListener;
-import dev.rico.server.remoting.event.Topic;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.core.ITopic;
 import org.apiguardian.api.API;
