+++ remoting/rico-remoting-server-spring-test/src/test/java/dev/rico/server/remoting/test/TestController.java
@@ -16,14 +16,14 @@
  */
 package dev.rico.server.remoting.test;
 
+import dev.rico.event.MessageEvent;
+import dev.rico.event.MessageListener;
+import dev.rico.event.Topic;
 import dev.rico.remoting.BeanManager;
 import dev.rico.server.remoting.RemotingAction;
 import dev.rico.server.remoting.RemotingController;
 import dev.rico.server.remoting.RemotingModel;
 import dev.rico.server.remoting.event.RemotingEventBus;
-import dev.rico.server.remoting.event.MessageEvent;
-import dev.rico.server.remoting.event.MessageListener;
-import dev.rico.server.remoting.event.Topic;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.annotation.PostConstruct;
