+++ remoting/rico-remoting-server-spring-test/src/test/java/dev/rico/server/remoting/test/eventbus/EventBusTestSubscriberController.java
@@ -16,11 +16,11 @@
  */
 package dev.rico.server.remoting.test.eventbus;
 
+import dev.rico.event.MessageEvent;
+import dev.rico.event.MessageListener;
 import dev.rico.server.remoting.RemotingController;
 import dev.rico.server.remoting.RemotingModel;
 import dev.rico.server.remoting.event.RemotingEventBus;
-import dev.rico.server.remoting.event.MessageEvent;
-import dev.rico.server.remoting.event.MessageListener;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.annotation.PostConstruct;
