+++ remoting/rico-remoting-server/src/main/java/dev/rico/internal/server/remoting/event/ListenerWithFilter.java
@@ -16,9 +16,9 @@
  */
 package dev.rico.internal.server.remoting.event;
 
+import dev.rico.event.MessageEventContext;
+import dev.rico.event.MessageListener;
 import dev.rico.internal.core.Assert;
-import dev.rico.server.remoting.event.MessageEventContext;
-import dev.rico.server.remoting.event.MessageListener;
 import org.apiguardian.api.API;
 
 import java.io.Serializable;
